feat: Loader loads images on demand from file and gives them to the boys
feat: The loader can now load images one batch at a time.
doc: The loader is fully documented.

It is very easy to use and I'm proud of it. It is my beloved child and I care for him very much.

To see how easy to use it is, see the file test.py. You can also run this very script and be in awe of my unparalleled prowess.