feat: Loader now runs in parallel
Loader is now its own process. Basically, in the past, it would block the main and interact in weird ways with its main loop. But now, the app is free of such burdens, and can decide everything from the main. Still need to adapt stuff though, but the horrible problem (main part) is now solved. You can take a big breath of relief if you want, that's what i also did.