feat: Hid more configloading in utility classes
I hate leaving the ConfigLoader synthax in the middle of my code. It's big, bulky, and takes freaking ages to write. And worse: if one day we want to change the structure of the config file, we will have to go through every single page of code to change every call of ConfigLoader. You would hate that, wouldn't you?
Well my good friend, today is your lucky day!
With the implementation of utility classes, I hide these calls in functions that are all in one place: the utility package.
From this commit onwards, calls for processes and loader config are in Utility classes.